pyproject.toml
[tool.poetry]
name = "fitness_fastapi"
version = "0.1.0"
description = ""
authors = ["sky-sky-code"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.1"
uvicorn = "^0.17.6"
aiohttp = "^3.8.1"
aerich = "^0.6.2"
asyncpg = "^0.25.0"
tortoise-orm = "^0.18.1"
django-environ = "^0.8.1"

[tool.poetry.dev-dependencies]


[tool.aerich]
tortoise_orm = "settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
